Fix test, extract extra-args.ts, simplify encodeAddressToAny func
