+++ ccip-sdk/src/ton/send.ts
@@ -1,20 +1,11 @@
-import { type Cell, Builder, beginCell, toNano } from '@ton/core'
+import { type Cell, beginCell, toNano } from '@ton/core'
 import { type TonClient, Address } from '@ton/ton'
-import { toBigInt } from 'ethers'
 
+import { CCIPError, CCIPErrorCode } from '../errors/index.ts'
+import type { AnyMessage, WithLogger } from '../types.ts'
+import { bytesToBuffer, encodeAddressToAny } from '../utils.ts'
+import { encodeExtraArgsCell } from './extra-args.ts'
 import type { UnsignedTONTx } from './types.ts'
-import { asSnakedCell } from './utils.ts'
-import { CCIPError, CCIPErrorCode, CCIPExtraArgsInvalidError } from '../errors/index.ts'
-import {
-  type ExtraArgs,
-  type SVMExtraArgsV1,
-  type SuiExtraArgsV1,
-  EVMExtraArgsV2Tag,
-  SVMExtraArgsV1Tag,
-  SuiExtraArgsV1Tag,
-} from '../extra-args.ts'
-import { type AnyMessage, type WithLogger, ChainFamily } from '../types.ts'
-import { bigIntReplacer, bytesToBuffer, encodeAddressToAny, getAddressBytes } from '../utils.ts'
 
 /** Opcode for Router ccipSend operation */
 export const CCIP_SEND_OPCODE = 0x31768d95
@@ -50,134 +41,6 @@ function encodeTokenAmounts(
       beginCell().storeAddress(Address.parse(ta.token)).storeUint(ta.amount, 256).endCell(),
     )
   }
-  return builder.endCell()
-}
-
-/**
- * Checks if extraArgs is SVMExtraArgsV1 format.
- */
-function isSVMExtraArgs(extraArgs: ExtraArgs): extraArgs is SVMExtraArgsV1 {
-  return 'computeUnits' in extraArgs
-}
-
-/**
- * Checks if extraArgs is SuiExtraArgsV1 format.
- */
-function isSuiExtraArgs(extraArgs: ExtraArgs): extraArgs is SuiExtraArgsV1 {
-  return 'receiverObjectIds' in extraArgs
-}
-
-/**
- * Encodes extraArgs as a Cell.
- *
- * Supports three formats based on the destination chain:
- * - GenericExtraArgsV2 (EVMExtraArgsV2) for EVM/TON/Aptos destinations
- * - SVMExtraArgsV1 for Solana destinations
- * - SuiExtraArgsV1 for Sui destinations
- *
- * @param extraArgs - Extra arguments for CCIP message
- * @returns Cell encoding the extra arguments
- * @throws {@link CCIPExtraArgsInvalidError} if extraArgs format is invalid
- */
-export function encodeExtraArgsCell(extraArgs: ExtraArgs): Cell {
-  if (isSVMExtraArgs(extraArgs)) {
-    return encodeSVMExtraArgsCell(extraArgs)
-  }
-  if (isSuiExtraArgs(extraArgs)) {
-    return encodeSuiExtraArgsCell(extraArgs)
-  }
-  return encodeEVMExtraArgsCell(extraArgs)
-}
-
-/**
- * Encodes extraArgs as a Cell using the GenericExtraArgsV2 (EVMExtraArgsV2) format.
- *
- * Format per chainlink-ton TL-B:
- * - tag: 32-bit opcode (0x181dcf10)
- * - gasLimit: Maybe<uint256> (1 bit flag + 256 bits if present)
- * - allowOutOfOrderExecution: 1 bit
- */
-function encodeEVMExtraArgsCell(extraArgs: ExtraArgs): Cell {
-  if (
-    Object.keys(extraArgs).filter((k) => k !== '_tag').length !== 2 ||
-    !('gasLimit' in extraArgs && 'allowOutOfOrderExecution' in extraArgs)
-  )
-    throw new CCIPExtraArgsInvalidError(ChainFamily.TON, JSON.stringify(extraArgs, bigIntReplacer))
-
-  let gasLimit: bigint | null = null
-  if (extraArgs.gasLimit > 0n) {
-    gasLimit = extraArgs.gasLimit
-  }
-
-  const builder = beginCell().storeUint(Number(EVMExtraArgsV2Tag), 32) // 0x181dcf10
-  builder.storeMaybeUint(gasLimit, 256)
-  builder.storeBit(extraArgs.allowOutOfOrderExecution)
-
-  return builder.endCell()
-}
-
-/**
- * Encodes extraArgs as a Cell using the SVMExtraArgsV1 format.
- *
- * Format per chainlink-ton TL-B:
- * - tag: 32-bit opcode (0x1f3b3aba)
- * - computeUnits: uint32
- * - accountIsWritableBitmap: uint64
- * - allowOutOfOrderExecution: bool
- * - tokenReceiver: uint256
- * - accounts: SnakedCell<uint256>
- */
-
-function encodeSVMExtraArgsCell(extraArgs: SVMExtraArgsV1): Cell {
-  // Encode accounts as a snaked cell of uint256 values
-  const accountsCell = asSnakedCell(extraArgs.accounts, (account: string) =>
-    new Builder().storeUint(toBigInt(getAddressBytes(account)), 256),
-  )
-
-  // Encode tokenReceiver as uint256
-  const tokenReceiver = extraArgs.tokenReceiver
-    ? toBigInt(getAddressBytes(extraArgs.tokenReceiver))
-    : 0n
-
-  const builder = beginCell()
-    .storeUint(Number(SVMExtraArgsV1Tag), 32) // 0x1f3b3aba
-    .storeUint(Number(extraArgs.computeUnits), 32)
-    .storeUint(extraArgs.accountIsWritableBitmap, 64)
-    .storeBit(extraArgs.allowOutOfOrderExecution)
-    .storeUint(tokenReceiver, 256) // uint256
-    .storeRef(accountsCell) // SnakedCell<uint256>
-
-  return builder.endCell()
-}
-
-/**
- * Encodes extraArgs as a Cell using the SuiExtraArgsV1 format.
- *
- * Format per chainlink-ton TL-B:
- * - tag: 32-bit opcode (0x21ea4ca9)
- * - gasLimit: uint256
- * - allowOutOfOrderExecution: bool
- * - tokenReceiver: uint256
- * - receiverObjectIds: SnakedCell<uint256>
- */
-function encodeSuiExtraArgsCell(extraArgs: SuiExtraArgsV1): Cell {
-  // Encode receiverObjectIds as a snaked cell of uint256 values
-  const objectIdsCell = asSnakedCell(extraArgs.receiverObjectIds, (objectId: string) =>
-    new Builder().storeUint(toBigInt(getAddressBytes(objectId)), 256),
-  )
-
-  // Encode tokenReceiver as uint256
-  const tokenReceiver = extraArgs.tokenReceiver
-    ? toBigInt(getAddressBytes(extraArgs.tokenReceiver))
-    : 0n
-
-  const builder = beginCell()
-    .storeUint(Number(SuiExtraArgsV1Tag), 32) // 0x21ea4ca9
-    .storeUint(extraArgs.gasLimit, 256)
-    .storeBit(extraArgs.allowOutOfOrderExecution)
-    .storeUint(tokenReceiver, 256) // uint256
-    .storeRef(objectIdsCell) // SnakedCell<uint256>
-
   return builder.endCell()
 }
